# Improve performance of Coordinates