In current implementation, `Coordinates` class internally update quaternion.
This process takes a lot of time.
This PR modified the process. 

```
import contextlib
import time

import numpy as np

import skrobot


@contextlib.contextmanager
def measure(name, log_enable=True):
    import time
    s = time.time()
    yield
    e = time.time()
    if log_enable:
        print("{}: {:.4f}sec".format(name, e - s))


Ns = [1, 10, 100, 1000, 10000]

for N in Ns:
    print(N)
    np.random.seed(0)
    with measure('coordinates.rotate'):
        for _ in range(N):
            radian = np.random.uniform(0, np.pi)
            skrobot.coordinates.Coordinates().rotate(
                radian, 'x')


for N in Ns:
    print(N)
    np.random.seed(0)
    with measure('coordinates.transformation'):
        for _ in range(N):
            a = skrobot.coordinates.base.random_coords()
            b = skrobot.coordinates.base.random_coords()
            a.transformation(b)
```
### Without PR
```
1
coordinates.rotate: 0.0198sec
10
coordinates.rotate: 0.0013sec
100
coordinates.rotate: 0.0125sec
1000
coordinates.rotate: 0.1408sec
10000
coordinates.rotate: 1.4056sec
1
coordinates.transformation: 0.0006sec
10
coordinates.transformation: 0.0042sec
100
coordinates.transformation: 0.0430sec
1000
coordinates.transformation: 0.4311sec
10000
coordinates.transformation: 4.7880sec
```

### With this PR
```
1
coordinates.rotate: 0.0212sec
10
coordinates.rotate: 0.0008sec
100
coordinates.rotate: 0.0055sec
1000
coordinates.rotate: 0.0648sec
10000
coordinates.rotate: 0.6726sec
1
coordinates.transformation: 0.0004sec
10
coordinates.transformation: 0.0024sec
100
coordinates.transformation: 0.0240sec
1000
coordinates.transformation: 0.2446sec
10000
coordinates.transformation: 2.4386sec
```